backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
djangorestframework-simplejwt==5.3.0
